Move ffmpeg_copy into ffmpeg_tools.py where it belongs
